feat: refactor search functionality into SearchClient and SearchInner components
